app/requirements.txt
flask==3.0.3
opencv-python-headless==4.10.0.84
numpy==2.1.1
PyYAML==6.0.2
python-telegram-bot==22.0
requests==2.32.3
paho-mqtt==2.1.0
